third_party/WebKit/LayoutTests/dom/attr/update-attribute-node-no-crash-expected.txt

Updating value of replaced attribute node

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS replaced = element.setAttributeNode(attr); replaced.name is "style"
PASS replaced.value is "background-color: red"
PASS attr.name is "style"
PASS attr.value is ""
PASS element.getAttribute('style') is ""
PASS element.getAttribute('StyLE') is ""
PASS attr.value = 'background-color: green'; attr.value is "background-color: green"
PASS element.getAttribute('stylE') is "background-color: green"
PASS successfullyParsed is true

TEST COMPLETE
